Update README.md
Coming to terms: the article is about a lab, so its very descriptive in what it is saying. For the common reader, I think it could be too detailed unless they were looking specifically for how to do something. 
Forwarding: very specific
Countering: kind of chunky and wordy
Looking ahead: I would change how it is formatted. Maybe make a table for the definition of popular foods. Be more concise on the methods so readers do not get lost in the jumble. Make it more visually appealing so readers aren't just staring about a paper reading mundane things about my soy lab. 
@matthewjmiller